% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIBrainCluster.R
\docType{class}
\name{ARIBrainCluster-class}
\alias{ARIBrainCluster-class}
\title{Class "ARIBrainCluster-class" for storing neuroimaging results of adaptive thresholding algorithm}
\value{
Returns an \code{\link{ARIBrainCluster-class}} object.
}
\description{
The \code{ARIBrainCluster-class} class is the output of a call to \code{\link{ARIBrainCluster}}. It inherits from the base class \code{ARICluster-class}, with the additional slots shown below, and stores the information needed for the next answering-query step, where all maximal supra-threshold clusters can be quickly found given a TDP threshold.
}
\section{Slots}{

\describe{
\item{\code{dims}}{Object of class "integer". Stores the 3D image dimensions.}

\item{\code{indexp}}{Object of class "integer". Stores the 3D in-mask voxel indices in the original order.}
}}

\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
