% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIBrainCluster.R
\name{ARIBrainCluster}
\alias{ARIBrainCluster}
\title{All-resolutions inference for cluster thresholding in neuroimaging (ARIBrainCluster)}
\usage{
ARIBrainCluster(Pmap, mask, conn = 18, alpha = 0.05)
}
\arguments{
\item{Pmap}{3D array of p-values, or a (character) nifti file name.}

\item{mask}{3D numeric/logical array. Non-zero/TRUE values in the mask indicate voxels in the brain. Voxels with missing values are treated as out of the brain. Alternatively, it can be a (character) nifti file name. If \code{mask} is not specified, it is assumed that none of the voxels have to be excluded.}

\item{conn}{Connectivity criterion (numeric or character). Can be chosen among 6 (or "face"), 18 (or "edge"), and 26 (or "vertex") for face, edge, and vertex connectivity. The default is using edge connectivity.}

\item{alpha}{Significance level. \code{alpha = 0.05} by default.}
}
\value{
Returns an \code{\link{ARIBrainCluster-class}} object, based on the base class \code{\link{ARICluster-class}}.
}
\description{
\code{ARIBrainCluster} is specially designed for brain imaging data analysis.
}
\examples{

pvalue_name <- system.file("extdata", "pvalue.nii.gz", package = "ARIbrain")
mask_name <- system.file("extdata", "mask.nii.gz", package = "ARIbrain")

# create an ARIBrainCluster-class object
aricluster <- ARIBrainCluster(Pmap = pvalue_name, mask = mask_name)

}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
