% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARICluster.R
\docType{class}
\name{ARICluster-class}
\alias{ARICluster-class}
\title{Class "ARICluster-class" for storing results of adaptive thresholding algorithm}
\value{
Returns an \code{\link{ARICluster-class}} object.
}
\description{
The \code{ARICluster-class} class is the output of a call to \code{\link{ARICluster}}. It stores the information needed for the next answering-query step, where all maximal supra-threshold clusters can be quickly found given a TDP threshold.
}
\section{Slots}{

\describe{
\item{\code{m}}{Object of class "integer". Stores the total number of nodes.}

\item{\code{alpha}}{Object of class "numeric". Stores the significance level.}

\item{\code{p}}{Object of class "numeric". Stores input p-values in the original order.}

\item{\code{stcs}}{Object of class "integer". Stores the representative nodes for all admissible supra-threshold clusters in the ascending order of TDP bounds.}

\item{\code{sizes}}{Object of class "integer". Stores the subtree sizes (cluster extents) for all nodes in the ascending order of p-values.}

\item{\code{tdps}}{Object of class "numeric". Stores the TDP lower confidence bounds for all supra-threshold clusters in the ascending order of p-values.}

\item{\code{childs}}{Object of class "list". Stores a list of vectors of child nodes for all nodes in the ascending order of p-values.}

\item{\code{marks}}{Object of class "integer". Initializes a vector to mark nodes within found clusters.}
}}

\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
