% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARICluster-function.R
\name{ARICluster}
\alias{ARICluster}
\title{All-resolutions inference (ARI) for cluster thresholding}
\usage{
ARICluster(p, adj, alpha = 0.05)
}
\arguments{
\item{p}{A vector of p-values.}

\item{adj}{A list of neighbours (unsorted orders that start from 1) for all nodes.}

\item{alpha}{Significance level. \code{alpha = 0.05} by default.}
}
\value{
Returns an \code{\link{ARICluster-class}} object.
}
\description{
Based on ARI, the adaptive thresholding algorithm is employed to answer queries.
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
