% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARICluster.R
\name{findLMs}
\alias{findLMs}
\title{Find all local minima}
\usage{
findLMs(aricluster)
}
\arguments{
\item{aricluster}{An \code{\link{ARICluster-class}} or \code{\link{ARIBrainCluster-class}} object.}
}
\value{
Returns a vector of all local minima.
}
\description{
\code{findLMs} is a generic function used to search for all local minima of p-values.
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
