% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPClusters.R
\name{length.TDPClusters}
\alias{length.TDPClusters}
\title{Operations on cluster list}
\usage{
\S4method{length}{TDPClusters}(x)
}
\arguments{
\item{x}{Object of class "TDPClusters" or "TDPBrainClusters". Stores the \code{\link{TDPClusters}} or \code{\link{TDPBrainClusters}} object, usually, a result of a call to \code{\link{TDPQuery}} or \code{\link{TDPChange}}.}
}
\description{
\code{length} method for class \code{\link{TDPClusters}} or \code{\link{TDPBrainClusters}}.
}
\details{
Some operations on \code{clusterlist} for class \code{\link{TDPClusters}} or \code{\link{TDPBrainClusters}}.
}
\examples{

# get number of clusters
length(tdpclusters)

# two ways to get all clusters:
tdpclusters[1:length(tdpclusters)]
tdpclusters@clusterlist

# get indices of the largest cluster
tdpclusters[[1]]

# get indices of the top 2 largest clusters
tdpclusters[1:2]

}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
