% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPBrainClusters.R
\name{summary.TDPBrainClusters}
\alias{summary.TDPBrainClusters}
\title{Summarize cluster information and generate summary table for found clusters.}
\usage{
summary(object, ..., rest = FALSE)
}
\description{
\code{summary} method for class \code{\link{TDPBrainClusters}}.
}
\details{
If the output is not assigned, the summary table will be printed on console.
}
\section{Slots}{

\describe{
\item{\code{object}}{Object of class "TDPBrainClusters". Stores the \code{\link{TDPBrainClusters}} object, usually, a result of a call to \code{\link{TDPQuery}}.}

\item{\code{rest}}{Object of class "logical". By default, \code{rest = FALSE} indicates that information on the rest of the brain will not be shown in the summary table.}
}}

\examples{

table <- summary(tdpclusters)      # get summary table (without rest)
summary(tdpclusters, rest = TRUE)  # print summary table (with rest)

}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
