% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPquery-class.R
\name{summaryCluster}
\alias{summaryCluster}
\title{Generate summary table for found clusters}
\usage{
summaryCluster(obj, object, rest = FALSE)
}
\arguments{
\item{obj}{An \code{\link{ARIBrainCluster-class}} object.}

\item{object}{A \code{\link{TDPquery-class}} object.}

\item{rest}{Logical; if \code{FALSE} (by default), information on the rest of the brain is not shown in the summary table.}
}
\value{
Returns a summary table reporting Size, FalseNull, TrueNull, ActiveProp and other summary statistics for each found cluster.
}
\description{
It is defined for \code{\link{ARIBrainCluster-class}} & \code{\link{TDPquery-class}} objects to create the summary table. If the output is not assigned, the summary table will be printed on console..
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
