% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARICluster.R
\name{summaryClusters}
\alias{summaryClusters}
\title{Generate summary table for found clusters}
\usage{
summaryCluster(aricluster, tdpclusters, rest = FALSE)
}
\arguments{
\item{aricluster}{An \code{\link{ARICluster-class}} or \code{\link{ARIBrainCluster-class}} object.}

\item{tdpclusters}{A \code{\link{TDPClusters}} object.}

\item{rest}{Logical; if \code{FALSE} (by default), information on the rest of the brain will not be shown in the summary table.}
}
\value{
Returns a summary table reporting Size, TDN (lower bound), #TrueNull (upper bound), TDP (lower bound), maximum statistic and the corresponding ID for each found cluster.
}
\description{
\code{summaryClusters} is a generic function used to create the summary table for all found clusters. The function invokes particular methods that depend on the class of the first argument.
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
