% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPBrainClusters.R
\docType{class}
\name{TDPBrainClusters}
\alias{TDPBrainClusters}
\title{Class "TDPBrainClusters" for outputting cluster information}
\value{
Returns a \code{\link{TDPBrainClusters}} object.
}
\description{
The class \code{TDPBrainClusters} is the output of a call to \code{\link{TDPQuery}}. It stores the resulting cluster information.
}
\section{Slots}{

\describe{
\item{\code{aricluster}}{Object of class "ARIBrainCluster-class". Stores the \code{\link{ARIBrainCluster-class}} object, usually, a result of a call to \code{\link{ARIBrainCluster}}.}

\item{\code{dims}}{Object of class "integer". Stores the dimensions of the input p-values.}

\item{\code{threshold}}{Object of class "numeric". Stores the TDP threshold.}

\item{\code{clusterlist}}{Object of class "list". Stores a list of found clusters, each including indices of nodes within that cluster. Here, the node index starts from 0. Please use \code{aricluster@indexp[tdpclusters@clusterlist[[i]]+1]} to access voxel indices in 3D space for the ith largest cluster.}
}}

\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
