% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPClusters.R
\name{TDPChange}
\alias{TDPChange}
\title{Change the size of a chosen cluster based on the TDP change inquiry}
\usage{
TDPChange(object, v, tdpchg=0.01)
}
\arguments{
\item{object}{Object of class "TDPClusters" or "TDPBrainClusters". Stores a \code{\link{TDPClusters}} or \code{\link{TDPBrainClusters}} object, usually, a result of a call to \code{\link{TDPQuery}} or \code{\link{TDPChange}}.}

\item{v}{Object of class "numeric". Stores a node index or a vector of 3D coordinates (one-based indexing), which is used to select a cluster in the cluster list saved in \code{object}.}

\item{tdpchg}{Object of class "numeric". Stores an expected change in the TDP bound, which should be within the range of \eqn{(-1,0) \cup (0,1)}. A positive value indicates the TDP bound is increased (i.e., cluster size is decreased), and a negative value indicates the TDP bound is decreased (i.e., cluster size is increased). \code{tdpchg = 0.01} by default.}
}
\value{
Returns a \code{\link{TDPClusters}} or \code{\link{TDPBrainClusters}} object of cluster list.
}
\description{
\code{TDPChange} is a generic function used to enlarge or shrink a cluster in terms of decreasing or increasing the TDP bound.
}
\examples{

pvalue_name <- system.file("extdata", "pvalue.nii.gz", package = "ARIbrain")
mask_name <- system.file("extdata", "mask.nii.gz", package = "ARIbrain")

# (1) create an ARIBrainCluster-class object
aricluster <- ARIBrainCluster(Pmap = pvalue_name, mask = mask_name)

# (2) answer query: find all maximal clusters given a TDP threshold
tdpclusters <- TDPQuery(aricluster, threshold = 0.7)
summary(tdpclusters)

# (3) change query: change the size of chosen cluster based on request
tdpchanges <- TDPChange(tdpclusters, v = 1, tdpchg =  0.01)  # increase TDP bound / decrease cluster size
tdpchanges <- TDPChange(tdpchanges,  v = 1, tdpchg =  0.01)  # further update the chosen cluster
summary(tdpchanges)

tdpchanges <- TDPChange(tdpchanges,  v = 1, tdpchg = -0.01)  # decrease TDP bound / increase cluster size
summary(tdpchanges)
    
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
