% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPquery-class.R
\docType{class}
\name{TDPquery-class}
\alias{TDPquery-class}
\title{Class "TDPquery" for outputting cluster information}
\value{
Returns an \code{\link{TDPquery-class}} object.
}
\description{
The class \code{TDPquery} is the output of a call to \code{\link{TDPquery}}. It stores the resulting cluster information.
}
\section{Slots}{

\describe{
\item{\code{gamma}}{Object of class "numeric". Stores the TDP threshold.}

\item{\code{clusterlist}}{Object of class "list". Stores a list of found clusters, each including sorting ranks for voxels within that cluster.}
}}

\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
