% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPquery-class.R
\name{writeCluster}
\alias{writeCluster}
\title{Write out result cluster image}
\usage{
writeCluster(obj, object, file = "aribrain.nii.gz", template = NULL)
}
\arguments{
\item{obj}{An \code{\link{ARIBrainCluster-class}} object.}

\item{object}{A \code{\link{TDPquery-class}} object.}

\item{file}{Character; The file name \code{"aribrain.nii.gz"} is used by default, where a file name that ends in .gz will be gzipped.}

\item{template}{A template object for writing Nifti image (please refer to \code{template} of \code{\link[RNifti]{writeNifti}} for more details).}
}
\description{
It is defined for \code{\link{ARIBrainCluster-class}} & \code{\link{TDPquery-class}} objects to write cluster image in Nifti format.
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
