% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIBrainCluster.R
\name{writeClusters}
\alias{writeClusters}
\title{Write out cluster image}
\usage{
writeCluster(tdpclusters, file, template, ...)
}
\arguments{
\item{tdpclusters}{A \code{\link{TDPBrainClusters}} object, usually, a result of a call to \code{\link{TDPQuery}} or \code{\link{TDPChange}}.}

\item{file}{Character; The file name for outputting cluster image, where a file name that ends in .gz will be gzipped.}

\item{template}{A template object for writing Nifti image (please refer to \code{template} of \code{\link[RNifti]{writeNifti}} for more details).}
}
\description{
\code{writeClusters} is defined for \code{\link{TDPBrainClusters}} object to write cluster image in Nifti format.
}
\details{
\code{writeClusters} works similar to \code{\link[RNifti]{writeNifti}}. All its parameters but the first are relayed to \code{\link[RNifti]{writeNifti}}.
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
