% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIbrain-package.R
\docType{_PACKAGE}
\name{ARIbrain-package}
\alias{ARIbrain-package}
\title{All-Resolutions Inference}
\description{
It performs All-Resolutions Inference (ARI) on fMRI data. As a main feature, it estimates lower bounds for the proportion of active voxels (true discovery proportion, TDP) in a set of clusters as, for example, given by a cluster-wise analysis. Additionally, given certain TDP thresholds, ARICluster analysis can quickly identify maximal clusters using ARI and an efficient adaptive cluster thresholding algorithm.
}
\examples{

# (1) ARI analysis
pvalue_name <- system.file("extdata", "pvalue.nii.gz", package="ARIbrain")
cluster_name <- system.file("extdata", "cluster_th_3.2.nii.gz", package="ARIbrain")
zstat_name <- system.file("extdata", "zstat.nii.gz", package="ARIbrain")
mask_name <- system.file("extdata", "mask.nii.gz", package="ARIbrain")

ARI(Pmap = pvalue_name, clusters= cluster_name, 
    mask=mask_name, Statmap = zstat_name)

# (2) ARICluster analysis
pvalue_name <- system.file("extdata", "pvalue.nii.gz", package = "ARIbrain")
mask_name <- system.file("extdata", "mask.nii.gz", package = "ARIbrain")

# create an ARIBrainCluster-class object
aricluster <- ARIBrainCluster(Pmap = pvalue_name, mask = mask_name)

# answer query: find all maximal clusters given a TDP threshold
tdpclusters <- TDPQuery(aricluster, threshold = 0.7)
summary(tdpclusters)

# change query: change the size of chosen cluster based on request
tdpchanges <- TDPChange(tdpclusters, v = 1, tdpchg =  0.01)  # increase TDP bound / decrease cluster size
tdpchanges <- TDPChange(tdpchanges,  v = 1, tdpchg =  0.01)  # further update the chosen cluster
summary(tdpchanges)

tdpchanges <- TDPChange(tdpchanges,  v = 1, tdpchg = -0.01)  # decrease TDP bound / increase cluster size
summary(tdpchanges)

}
\author{
all of us
}
