% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIcluster-function.R
\name{ARIcluster}
\alias{ARIcluster}
\title{All-resolutions inference (ARI) for cluster thresholding}
\usage{
ARIcluster(Pmap, mask = NULL, conn = 18, alpha = 0.05)
}
\arguments{
\item{Pmap}{3D array of p-values or a (character) nifti file name.}

\item{mask}{3D array of logicals (i.e., \code{TRUE}/\code{FALSE} for in/out of the brain). Alternatively, it may be a (character) nifti file name. If \code{mask = NULL}, it is assumed that none of the voxels have to be excluded.}

\item{conn}{Connectivity criterion, which can be chosen among 6, 18 (the default), and 26 for face, edge, and vertex connectivity.}

\item{alpha}{Significance level. \code{alpha = 0.05} by default.}
}
\value{
Returns an \code{\link{ARIcluster-class}} object.
}
\description{
Based on ARI, the adaptive thresholding algorithm is employed to answer queries.
}
\examples{
pvalue_name <- system.file("extdata", "pvalue.nii.gz", package = "ARIbrain")
mask_name <- system.file("extdata", "mask.nii.gz", package = "ARIbrain")

print(pvalue_name)
print(mask_name)

# (1) create an ARIcluster object
ariclstr <- ARIcluster(Pmap = pvalue_name, mask = mask_name)
show(ariclstr)

# (2) answer queries: find all maximal clusters given a TDP threshold
res <- TDPquery(ariclstr, gamma = 0.7)
res

}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
