% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPquery-function.R
\name{TDPquery}
\alias{TDPquery}
\title{Answer queries with a TDP threshold}
\usage{
TDPquery(ARIBrainCluster, gamma)
}
\arguments{
\item{ARIBrainCluster}{An \code{\link{ARIBrainCluster-class}} object.}

\item{gamma}{A TDP threshold for forming maximal clusters.}
}
\value{
Returns a \code{\link{TDPquery-class}} object of cluster list.
}
\description{
Given a TDP threshold, it quickly finds all maximal supra-threshold clusters, each with the TDP not smaller than the threshold.
}
\examples{
pvalue_name <- system.file("extdata", "pvalue.nii.gz", package = "ARIbrain")
mask_name <- system.file("extdata", "mask.nii.gz", package = "ARIbrain")

# (1) create an ARIBrainCluster object
ari <- ARIBrainCluster(Pmap = pvalue_name, mask = mask_name)

# (2) answer queries: find all maximal clusters given a TDP threshold
res <- TDPquery(ari, gamma = 0.7)
res@clusterlist  # access cluster list (sorting ranks that start from 0)

# (3) summarize cluster information and write image file
summaryCluster(ari, res, rest = TRUE)
writeCluster(ari, res, file = "ari.nii.gz", template = mask_name)
    
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
