% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARICluster.R
\name{TDPQuery}
\alias{TDPQuery}
\title{Answer query with a TDP threshold}
\usage{
TDPQuery(aricluster, threshold)
}
\arguments{
\item{aricluster}{An \code{\link{ARICluster-class}} or \code{\link{ARIBrainCluster-class}} object.}

\item{threshold}{A TDP threshold for forming maximal clusters.}
}
\value{
Returns a \code{\link{TDPClusters}}or \code{\link{TDPBrainClusters}} object of cluster list.
}
\description{
\code{TDPQuery} is a generic function used to quickly find all maximal supra-threshold clusters given a TDP threshold, where each cluster has the TDP not smaller than the threshold. The function invokes particular methods that depend on the class of the first argument.
}
\examples{

pvalue_name <- system.file("extdata", "pvalue.nii.gz", package = "ARIbrain")
mask_name <- system.file("extdata", "mask.nii.gz", package = "ARIbrain")

# (1) create an ARIBrainCluster-class object
aricluster <- ARIBrainCluster(Pmap = pvalue_name, mask = mask_name)

# (2) answer query: find all maximal clusters given a TDP threshold
tdpclusters <- TDPQuery(aricluster, threshold = 0.7)
summary(tdpclusters)
    
}
\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
